Доработана возможность поставки Double, а не INT

+++ app/src/main/java/com/example/diningroomfactory/Models/FactoryProduct.kt
@@ -14,6 +14,6 @@ data class FactoryProduct(
     var productId: String = "",
 
     @PropertyName("productCount")
-    var productCount: Int = 0
+    var productCount: Double = 0.0
 
 )
